requirements.txt
Django==5.2.1
environs==14.1.1
aiogram==3.20.0
python-dotenv==1.0.1